FastAPI_Nginx_Minukube/requirements.txt
fastapi
uvicorn
msal
python-dotenv
